requirements.txt
pypdf
langchain~=0.2.0
langchain_community
chromadb # Vector storage
pytest
boto3
bs4~=0.0.2
ollama
unstructured
requests~=2.32.2
beautifulsoup4~=4.12.3
PyYAML~=6.0.1
tqdm
pillow
pandas
PyMuPDF